Extract icons component into App component.
It's a small component, so probably doesn't require its own component.

Rather than using `dangerouslySetInnerHTML` and `for(etc...)`, you can use `Array.map()` to iterate over an array.

I changed the shape of the config to make it less brittle, and it makes it simpler for the `.map` to extract the data.

Let me know if you have any questions, I'm happy to help :)